Bugfix: boost::format calls stream << for enum types, leading to infinite recursion when it is used in the stream handler.

+++ src/main.cpp
@@ -194,7 +194,7 @@ void AfIncomingMsg(std::shared_ptr<znp::ZnpApi> api,
                   boost::str(boost::format("%s%08X/%d/%s/%04X") % mqtt_prefix %
                              ieee_addr % (unsigned int)SrcEndpoint %
                              zcl::to_string((zcl::ZclClusterId)ClusterId) %
-                             std::get<0>(attribute));
+                             (unsigned int)std::get<0>(attribute));
               std::stringstream message_stream;
               message_stream << std::get<1>(attribute);
               publishes.push_back(
